requirements.txt
discord
python-dotenv
requests
discord-py-interactions
lib
sec-api
replicate